requirements_cuda118.txt
absl-py==1.3.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apex==0.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
audioread==3.0.0
azure-core==1.30.2
azure-storage-blob==12.22.0
backcall==0.2.0
bcrypt==4.2.0
beanie==1.26.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpathlib==0.18.1
cloudpickle==2.2.0
cmake==3.24.1.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.0.6
cryptography==43.0.0
cuda-python @ file:///rapids/cuda_python-11.7.0%2B0.g95a2041.dirty-cp38-cp38-linux_x86_64.whl
cudf @ file:///rapids/cudf-22.8.0a0%2B304.g6ca81bbc78.dirty-cp38-cp38-linux_x86_64.whl
cugraph @ file:///rapids/cugraph-22.8.0a0%2B132.g2daa31b6.dirty-cp38-cp38-linux_x86_64.whl
cuml @ file:///rapids/cuml-22.8.0a0%2B52.g73b8d00d0.dirty-cp38-cp38-linux_x86_64.whl
cupy-cuda118 @ file:///rapids/cupy_cuda118-11.0.0-cp38-cp38-linux_x86_64.whl
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
dask @ file:///rapids/dask-2022.7.1-py3-none-any.whl
dask-cuda @ file:///rapids/dask_cuda-22.8.0a0%2B36.g9860cad-py3-none-any.whl
dask-cudf @ file:///rapids/dask_cudf-22.8.0a0%2B304.g6ca81bbc78.dirty-py3-none-any.whl
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distributed @ file:///rapids/distributed-2022.7.1-py3-none-any.whl
distro==1.9.0
dlib==19.24.6
dnspython==2.6.1
ecdsa==0.19.0
elevenlabs==1.7.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.0.4
execnet==1.9.0
executing==1.2.0
expecttest==0.1.3
fastapi==0.112.2
fastapi-cli==0.0.5
fastjsonschema==2.16.2
fastrlock==0.8.1
flatbuffers==24.3.25
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2022.11.0
functorch==1.13.0a0+936e930
google-auth==2.14.1
google-auth-oauthlib==0.4.6
graphsurgeon @ file:///workspace/TensorRT-8.5.1.7/graphsurgeon/graphsurgeon-0.4.6-py2.py3-none-any.whl
greenlet==3.0.3
grpcio==1.50.0
h11==0.14.0
HeapDict==1.0.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
hypothesis==5.35.1
idna==3.4
imageio==2.35.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
intel-openmp==2021.4.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.2.0
json5==0.9.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.0
jupyter-tensorboard @ git+https://github.com/cliffwoolley/jupyter_tensorboard.git@ffa7e26138b82549453306e06b535a9ac36db17a
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==2.3.2
jupyterlab-pygments==0.2.2
jupyterlab-server==1.2.0
jupytext==1.14.1
kiwisolver==1.4.4
langchain==0.2.14
langchain-core==0.2.35
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langcodes==3.3.0
langsmith==0.1.104
lazy-model==0.2.0
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
mistune==2.0.4
mkl==2021.1.1
mkl-devel==2021.1.1
mkl-include==2021.1.1
mock==4.0.3
motor==3.5.1
mpmath==1.3.0
msgpack==1.0.4
multidict==6.0.5
murmurhash==1.0.9
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.4.10
numba==0.56.4
numpy==1.22.2
nvidia-dali-cuda110==1.18.0
nvidia-pyindex==1.0.9
nvtx==0.2.5
oauthlib==3.2.2
onnx @ file:///opt/pytorch/pytorch/third_party/onnx
onnxruntime==1.19.0
openai==1.42.0
opencv @ file:///opencv-4.6.0/modules/python/package
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
passlib==1.7.4
pathy==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
pluggy==1.0.0
polygraphy==0.42.1
pooch==1.6.0
preshed==3.0.8
prettytable==3.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow @ file:///rapids/pyarrow-8.0.0-cp38-cp38-linux_x86_64.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pycocotools @ git+https://github.com/nvidia/cocoapi.git@8b8fd68576675c3ee77402e61672d65a7d826ddf#subdirectory=PythonAPI
pycparser==2.21
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.13.0
pylibcugraph @ file:///rapids/pylibcugraph-22.8.0a0%2B132.g2daa31b6.dirty-cp38-cp38-linux_x86_64.whl
PyMatting==1.1.12
pymongo==4.8.0
pynvml==11.4.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-rerunfailures==10.2
pytest-shard==0.1.2
pytest-xdist==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-hostlist==1.22
python-jose==3.3.0
python-multipart==0.0.9
pytorch-quantization==2.1.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
raft @ file:///rapids/raft-22.8.0a0%2B70.g9070c30.dirty-cp38-cp38-linux_x86_64.whl
regex==2022.10.31
rembg==2.0.59
replicate==0.32.0
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.8.0
rmm @ file:///rapids/rmm-22.8.0a0%2B62.gf6bf047.dirty-cp38-cp38-linux_x86_64.whl
rsa==4.9
scikit-image==0.21.0
scikit-learn @ file:///rapids/scikit_learn-0.24.2-cp38-cp38-manylinux2010_x86_64.whl
scipy==1.10.1
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
smart-open==5.2.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.3
sphinx-glpi-theme==0.3
SQLAlchemy==2.0.32
srsly==2.4.5
stack-data==0.6.1
starlette==0.38.2
sympy==1.13.2
tbb==2021.7.1
tblib==1.7.0
tenacity==8.5.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorrt @ file:///workspace/TensorRT-8.5.1.7/python/tensorrt-8.5.1.7-cp38-none-linux_x86_64.whl
terminado==0.17.0
thinc==8.2.5
threadpoolctl==3.1.0
tifffile==2023.7.10
tiktoken==0.7.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.0a0+936e930
torch-tensorrt @ file:///opt/pytorch/torch_tensorrt/py/dist/torch_tensorrt-1.3.0a0-cp38-cp38-linux_x86_64.whl
torchtext @ git+https://github.com/pytorch/text@fae8e8cabf7adcbbc2f09c0520216288fd53f33b
torchvision @ file:///opt/pytorch/vision
tornado==6.1
tqdm==4.64.1
traitlets==5.5.0
transformer-engine @ git+https://github.com/NVIDIA/TransformerEngine.git@29e9e278a331eeba8b44d84a718a4bc49a50dd5b
treelite @ file:///rapids/treelite-2.4.0-py3-none-manylinux2014_x86_64.whl
treelite-runtime @ file:///rapids/treelite_runtime-2.4.0-py3-none-manylinux2014_x86_64.whl
typer==0.12.5
typing_extensions==4.12.2
ucx-py @ file:///rapids/ucx_py-0.27.0a0%2B29.ge9e81f8-cp38-cp38-linux_x86_64.whl
uff @ file:///workspace/TensorRT-8.5.1.7/uff/uff-0.6.9-py2.py3-none-any.whl
urllib3==1.26.12
uvicorn==0.30.6
uvloop==0.20.0
wasabi==0.10.1
watchfiles==0.23.0
wcwidth==0.2.5
weasel==0.4.1
webencodings==0.5.1
websockets==13.0
Werkzeug==2.2.2
xdoctest==1.0.2
xgboost @ file:///rapids/xgboost-1.6.1-cp38-cp38-linux_x86_64.whl
yarl==1.9.4
zict==2.2.0
zipp==3.10.0
